src/kubedoor-master/requirements.txt
loguru
asyncio
aiohttp
requests
clickhouse_driver
pydantic
mysql-connector-python
PyYAML
typing-extensions
python-dotenv
clickhouse-connect
huaweicloudsdkcore==3.1.165
huaweicloudsdkswr==3.1.165
aliyun-python-sdk-core==2.16.0
aliyun-python-sdk-cr==4.1.2